+++ src/shared.rs
@@ -1,6 +1,5 @@
 pub mod backend;
 pub mod client_id;
-pub mod common_conditions;
 pub mod event;
 pub mod protocol;
 pub mod replication;
@@ -47,7 +46,7 @@ pub struct RepliconSharedPlugin {
     .add_server_trigger::<ProtocolMismatch>(Channel::Unreliable)
     .make_trigger_independent::<ProtocolMismatch>() // Let client receive it without replication.
     .add_observer(start_game)
-    .add_systems(Update, send_info.run_if(client_just_connected));
+    .add_systems(OnEnter(ClientState::Connected), send_info);
 
     fn send_info(
         mut commands: Commands,
@@ -135,6 +134,8 @@ impl Plugin for RepliconSharedPlugin {
             .register_type::<ConnectedClient>()
             .register_type::<NetworkIdMap>()
             .register_type::<NetworkStats>()
+            .init_state::<ClientState>()
+            .init_state::<ServerState>()
             .init_resource::<ProtocolHasher>()
             .init_resource::<NetworkIdMap>()
             .init_resource::<TrackMutateMessages>()